stage2/requirements.txt
accelerate==0.21.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-cors==0.8.1
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
async-timeout==5.0.1
attrs==25.4.0
bitsandbytes==0.45.5
certifi==2025.6.15
cffi==2.0.0
charset-normalizer==3.4.4
click==8.2.1
cloudpickle==3.1.2
codetiming==1.4.0
colorful==0.5.8
contourpy==1.3.2
cryptography==46.0.4
cuda-bindings==12.9.4
cuda-pathfinder==1.3.3
cycler==0.12.1
datasets==4.5.0
deepspeed==0.12.6
dill==0.4.0
distlib==0.4.0
distro==1.9.0
einops==0.6.1
einops-exts==0.0.4
exceptiongroup==1.3.1
fastapi==0.115.13
ffmpy==0.6.0
filelock==3.20.3
flash-attn==2.7.3
fonttools==4.58.4
frozenlist==1.8.0
fsspec==2025.10.0
ftfy==6.3.1
fvcore==0.1.5.post20221221
gitdb==4.0.12
GitPython==3.1.46
google-api-core==2.29.0
google-auth==2.48.0
googleapis-common-protos==1.72.0
gradio==4.16.0
gradio_client==0.8.1
grpcio==1.76.0
h11==0.14.0
hf-xet==1.2.0
hjson==3.1.0
httpcore==0.17.3
httpx==0.24.0
huggingface-hub==0.33.1
hydra-core==1.3.2
idna==3.10
importlib_metadata==8.7.1
importlib_resources==6.5.2
iniconfig==2.1.0
iopath==0.1.10
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.8
latex2mathml==3.78.0
# Use local project source instead of machine-specific editable path
# (Install from this repo root with: pip install -e .)
loralib==0.1.2
markdown-it-py==3.0.0
markdown2==2.5.3
MarkupSafe==2.1.5
matplotlib==3.10.3
mdurl==0.1.2
mpi4py==4.1.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
multiprocess==0.70.18
narwhals==1.44.0
networkx==3.4.2
ninja==1.11.1.4
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
openai==1.107.3
opencensus==0.11.4
opencensus-context==0.1.3
openmpi==5.0.8
opentelemetry-api==1.39.1
opentelemetry-exporter-prometheus==0.60b1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.7
packaging==26.0
pandas==2.3.3
peft==0.10.0
pillow==10.4.0
platformdirs==4.5.1
pluggy==1.6.0
portalocker==3.2.0
prometheus_client==0.24.1
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.5
psutil==7.2.2
py-cpuinfo==9.0.0
py-spy==0.4.1
pyarrow==23.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pybind11==3.0.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
pylatexenc==2.10
pynvml==12.0.0
pyparsing==3.2.3
pytest==8.4.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
ray==2.53.0
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
rich==14.0.0
rpds-py==0.30.0
rsa==4.9.1
ruff==0.12.0
safetensors==0.7.0
scikit-learn==1.2.2
scipy==1.15.3
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==2.51.0
setproctitle==1.3.6
shellingham==1.5.4
shortuuid==1.0.13
six==1.17.0
smart_open==7.5.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.46.2
svgwrite==1.4.3
sympy==1.13.1
tabulate==0.9.0
tensordict==0.6.2
termcolor==3.3.0
threadpoolctl==3.6.0
timm==0.6.13
tokenizers==0.15.1
tomli==2.2.1
tomlkit==0.12.0
torch==2.1.2
torchaudio==2.1.2
torchdata==0.11.0
torchvision==0.16.2
tqdm==4.67.3
transformers==4.37.2
triton==2.1.0
typer==0.16.0
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.34.3
virtualenv==20.36.1
wandb==0.24.1
wavedrom==2.0.3.post3
wcwidth==0.2.13
websockets==11.0.3
wget==3.2
wrapt==2.1.1
xxhash==3.6.0
yacs==0.1.8
yarl==1.22.0
zipp==3.23.0
